requirements.txt
gdown
pandas
numpy
matplotlib
sentence-transformers
umap-learn
yt-dlp
